+++ pages/index.js
@@ -1,13 +1,45 @@
-import LatestPosts from '../components/blog/LatestPosts'
 import Layout from '../components/common/Layout'
+import Link from 'next/link'
+import PropTypes from 'prop-types'
+import fetch from 'isomorphic-unfetch'
 
-const Index = () => (
+const Homepage = ({posts}) => (
   <Layout>
     <section className="container">
       <img src={require('../public/images/wdscamp.jpg?resize&size=728')} />
-      <LatestPosts />
+      <aside className="latest-posts">
+        <h3>Latest from the blog</h3>
+        <ul>
+          {posts.map((post) => (
+            <li key={post.id}>
+              <Link href="/blog/[...all]" as={`/blog/${post.id}/${post.slug}`}>
+                <a dangerouslySetInnerHTML={{__html: post.title.rendered}} />
+              </Link>
+            </li>
+          ))}
+        </ul>
+      </aside>
     </section>
   </Layout>
 )
 
-export default Index
+export async function getStaticProps() {
+  const res = await fetch('https://webdevstudios.com/wp-json/wp/v2/posts')
+  const posts = await res.json()
+
+  return {
+    props: {
+      posts
+    }
+  }
+}
+
+Homepage.propTypes = {
+  posts: PropTypes.array
+}
+
+Homepage.defaultProps = {
+  posts: []
+}
+
+export default Homepage
